Modify plugin filtering fudge to only bother the test...
does make it wierdly slow though.
